Added CollectionInterface and BaseCollection for subclassing
