+++ src/Sulu/Bundle/MediaBundle/Entity/Collection.php
@@ -10,51 +10,14 @@
 
 namespace Sulu\Bundle\MediaBundle\Entity;
 
+use Doctrine\ORM\Mapping as ORM;
 use JMS\Serializer\Annotation\Exclude;
 
 /**
  * Collection
  */
-class Collection
+class Collection extends BaseCollection
 {
-    /**
-     * @var string
-     */
-    private $style;
-
-    /**
-     * @var integer
-     * @Exclude
-     */
-    private $lft;
-
-    /**
-     * @var integer
-     * @Exclude
-     */
-    private $rgt;
-
-    /**
-     * @var integer
-     * @Exclude
-     */
-    private $depth;
-
-    /**
-     * @var \DateTime
-     */
-    private $created;
-
-    /**
-     * @var \DateTime
-     */
-    private $changed;
-
-    /**
-     * @var integer
-     */
-    private $id;
-
     /**
      * @var \Doctrine\Common\Collections\Collection
      */
@@ -72,74 +35,6 @@ class Collection
     private $children;
 
     /**
-     * @var \Sulu\Bundle\MediaBundle\Entity\Collection
-     */
-    private $parent;
-
-    /**
-     * @var \Sulu\Bundle\MediaBundle\Entity\CollectionType
-     */
-    private $type;
-
-    /**
-     * @var \Sulu\Component\Security\UserInterface
-     * @Exclude
-     */
-    private $changer;
-
-    /**
-     * @var \Sulu\Component\Security\UserInterface
-     * @Exclude
-     */
-    private $creator;
-
-    /**
-     * Set changer
-     *
-     * @param \Sulu\Component\Security\UserInterface $changer
-     * @return Collection
-     */
-    public function setChanger(\Sulu\Component\Security\UserInterface $changer = null)
-    {
-        $this->changer = $changer;
-
-        return $this;
-    }
-
-    /**
-     * Get changer
-     *
-     * @return \Sulu\Component\Security\UserInterface
-     */
-    public function getChanger()
-    {
-        return $this->changer;
-    }
-
-    /**
-     * Set creator
-     *
-     * @param \Sulu\Component\Security\UserInterface $creator
-     * @return Collection
-     */
-    public function setCreator(\Sulu\Component\Security\UserInterface $creator = null)
-    {
-        $this->creator = $creator;
-
-        return $this;
-    }
-
-    /**
-     * Get creator
-     *
-     * @return \Sulu\Component\Security\UserInterface
-     */
-    public function getCreator()
-    {
-        return $this->creator;
-    }
-
-    /**
      * Constructor
      */
     public function __construct()
@@ -147,154 +42,6 @@ class Collection
         $this->meta = new \Doctrine\Common\Collections\ArrayCollection();
         $this->media = new \Doctrine\Common\Collections\ArrayCollection();
         $this->children = new \Doctrine\Common\Collections\ArrayCollection();
-    }
-
-    /**
-     * Set style
-     *
-     * @param string $style
-     * @return Collection
-     */
-    public function setStyle($style)
-    {
-        $this->style = $style;
-
-        return $this;
-    }
-
-    /**
-     * Get style
-     *
-     * @return string
-     */
-    public function getStyle()
-    {
-        return $this->style;
-    }
-
-    /**
-     * Set lft
-     *
-     * @param integer $lft
-     * @return Collection
-     */
-    public function setLft($lft)
-    {
-        $this->lft = $lft;
-
-        return $this;
-    }
-
-    /**
-     * Get lft
-     *
-     * @return integer
-     */
-    public function getLft()
-    {
-        return $this->lft;
-    }
-
-    /**
-     * Set rgt
-     *
-     * @param integer $rgt
-     * @return Collection
-     */
-    public function setRgt($rgt)
-    {
-        $this->rgt = $rgt;
-
-        return $this;
-    }
-
-    /**
-     * Get rgt
-     *
-     * @return integer
-     */
-    public function getRgt()
-    {
-        return $this->rgt;
-    }
-
-    /**
-     * Set depth
-     *
-     * @param integer $depth
-     * @return Collection
-     */
-    public function setDepth($depth)
-    {
-        $this->depth = $depth;
-
-        return $this;
-    }
-
-    /**
-     * Get depth
-     *
-     * @return integer
-     */
-    public function getDepth()
-    {
-        return $this->depth;
-    }
-
-    /**
-     * Set created
-     *
-     * @param \DateTime $created
-     * @return Collection
-     */
-    public function setCreated($created)
-    {
-        $this->created = $created;
-
-        return $this;
-    }
-
-    /**
-     * Get created
-     *
-     * @return \DateTime
-     */
-    public function getCreated()
-    {
-        return $this->created;
-    }
-
-    /**
-     * Set changed
-     *
-     * @param \DateTime $changed
-     * @return Collection
-     */
-    public function setChanged($changed)
-    {
-        $this->changed = $changed;
-
-        return $this;
-    }
-
-    /**
-     * Get changed
-     *
-     * @return \DateTime
-     */
-    public function getChanged()
-    {
-        return $this->changed;
-    }
-
-    /**
-     * Get id
-     *
-     * @return integer
-     */
-    public function getId()
-    {
-        return $this->id;
     }
 
     /**
@@ -356,7 +103,7 @@ class Collection
     /**
      * Get media
      *
-     * @return \Doctrine\Common\Collections\Collection
+     * @return \Doctrine\Common\Collections\CollectionInterface
      */
     public function getMedia()
     {
@@ -366,10 +113,10 @@ class Collection
     /**
      * Add children
      *
-     * @param \Sulu\Bundle\MediaBundle\Entity\Collection $children
+     * @param \Sulu\Bundle\MediaBundle\Entity\CollectionInterface $children
      * @return Collection
      */
-    public function addChildren(\Sulu\Bundle\MediaBundle\Entity\Collection $children)
+    public function addChildren(\Sulu\Bundle\MediaBundle\Entity\CollectionInterface $children)
     {
         $this->children[] = $children;
 
@@ -379,9 +126,9 @@ class Collection
     /**
      * Remove children
      *
-     * @param \Sulu\Bundle\MediaBundle\Entity\Collection $children
+     * @param \Sulu\Bundle\MediaBundle\Entity\CollectionInterface $children
      */
-    public function removeChildren(\Sulu\Bundle\MediaBundle\Entity\Collection $children)
+    public function removeChildren(\Sulu\Bundle\MediaBundle\Entity\CollectionInterface $children)
     {
         $this->children->removeElement($children);
     }
@@ -389,56 +136,10 @@ class Collection
     /**
      * Get children
      *
-     * @return \Doctrine\Common\Collections\Collection
+     * @return \Doctrine\Common\Collections\CollectionInterface
      */
     public function getChildren()
     {
         return $this->children;
-    }
-
-    /**
-     * Set parent
-     *
-     * @param \Sulu\Bundle\MediaBundle\Entity\Collection $parent
-     * @return Collection
-     */
-    public function setParent(\Sulu\Bundle\MediaBundle\Entity\Collection $parent = null)
-    {
-        $this->parent = $parent;
-
-        return $this;
-    }
-
-    /**
-     * Get parent
-     *
-     * @return \Sulu\Bundle\MediaBundle\Entity\Collection
-     */
-    public function getParent()
-    {
-        return $this->parent;
-    }
-
-    /**
-     * Set type
-     *
-     * @param \Sulu\Bundle\MediaBundle\Entity\CollectionType $type
-     * @return Collection
-     */
-    public function setType(\Sulu\Bundle\MediaBundle\Entity\CollectionType $type)
-    {
-        $this->type = $type;
-
-        return $this;
-    }
-
-    /**
-     * Get type
-     *
-     * @return \Sulu\Bundle\MediaBundle\Entity\CollectionType
-     */
-    public function getType()
-    {
-        return $this->type;
     }
 }
