+++ src/Sulu/Bundle/MediaBundle/Resources/config/doctrine/Collection.orm.xml
@@ -6,52 +6,13 @@
 
     <entity name="Sulu\Bundle\MediaBundle\Entity\Collection" table="me_collections" repository-class="Sulu\Bundle\MediaBundle\Entity\CollectionRepository">
 
-        <id name="id" type="integer" column="id">
-            <generator strategy="AUTO"/>
-        </id>
-        <field name="style" type="string" column="style" length="255" nullable="true"/>
-        <field name="lft" type="integer" column="lft">
-            <gedmo:tree-left/>
-        </field>
-        <field name="rgt" type="integer" column="rgt">
-            <gedmo:tree-right/>
-        </field>
-        <field name="depth" type="integer" column="depth">
-            <gedmo:tree-level/>
-        </field>
-        <field name="created" type="datetime" column="created"/>
-        <field name="changed" type="datetime" column="changed"/>
-
         <one-to-many field="meta" target-entity="Sulu\Bundle\MediaBundle\Entity\CollectionMeta" mapped-by="collection">
             <cascade>
                 <cascade-persist />
             </cascade>
         </one-to-many>
         <one-to-many field="media" target-entity="Sulu\Bundle\MediaBundle\Entity\Media" mapped-by="collection"/>
-        <one-to-many field="children" target-entity="Sulu\Bundle\MediaBundle\Entity\Collection" mapped-by="parent"/>
+        <one-to-many field="children" target-entity="Sulu\Bundle\MediaBundle\Entity\CollectionInterface" mapped-by="parent"/>
 
-        <many-to-one field="parent" target-entity="Sulu\Bundle\MediaBundle\Entity\Collection" inversed-by="children">
-            <join-columns>
-                <join-column name="idCollectionsParent" referenced-column-name="id" on-delete="SET NULL" />
-            </join-columns>
-            <gedmo:tree-parent/>
-        </many-to-one>
-        <many-to-one field="type" target-entity="Sulu\Bundle\MediaBundle\Entity\CollectionType" inversed-by="collections">
-            <join-columns>
-                <join-column name="idCollectionTypes" referenced-column-name="id" nullable="false"/>
-            </join-columns>
-        </many-to-one>
-        <many-to-one field="changer" target-entity="Sulu\Component\Security\UserInterface">
-            <join-columns>
-                <join-column name="idUsersChanger" referenced-column-name="id" on-delete="SET NULL" nullable="true"/>
-            </join-columns>
-        </many-to-one>
-        <many-to-one field="creator" target-entity="Sulu\Component\Security\UserInterface">
-            <join-columns>
-                <join-column name="idUsersCreator" referenced-column-name="id" on-delete="SET NULL" nullable="true"/>
-            </join-columns>
-        </many-to-one>
-
-        <gedmo:tree type="nested"/>
     </entity>
 </doctrine-mapping>
