+++ src/Sulu/Bundle/MediaBundle/Resources/config/doctrine/CollectionMeta.orm.xml
@@ -9,7 +9,7 @@
         <field name="title" type="string" column="title" length="255"/>
         <field name="description" type="text" column="description" nullable="true"/>
         <field name="locale" type="string" column="locale" length="5"/>
-        <many-to-one field="collection" target-entity="Sulu\Bundle\MediaBundle\Entity\Collection" inversed-by="meta">
+        <many-to-one field="collection" target-entity="Sulu\Bundle\MediaBundle\Entity\CollectionInterface" inversed-by="meta">
             <join-columns>
                 <join-column name="idCollections" referenced-column-name="id" nullable="false" on-delete="CASCADE"/>
             </join-columns>
